+++ packages/react-components/react-headless-components-preview/library/src/components/Badge/renderBadge.tsx
@@ -1,6 +1,1 @@
-import { renderBadge_unstable } from '@fluentui/react-badge';
-
-/**
- * Renders the final JSX of the Badge component, given the state.
- */
-export const renderBadge = renderBadge_unstable;
+export { renderBadge_unstable as renderBadge } from '@fluentui/react-badge';