+++ packages/react-components/react-migration-v0-v9/library/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-migration-v0-v9
 
-This log was last generated on Mon, 29 Jun 2026 15:15:07 GMT and should not be manually modified.
+This log was last generated on Wed, 15 Jul 2026 14:10:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.6.33](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v0-v9_v9.6.33)
+
+Wed, 15 Jul 2026 14:10:41 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-migration-v0-v9_v9.6.32..@fluentui/react-migration-v0-v9_v9.6.33)
+
+### Patches
+
+- Bump @fluentui/react-components to v9.74.4 ([PR #36401](https://github.com/microsoft/fluentui/pull/36401) by beachball)
 
 ## [9.6.32](https://github.com/microsoft/fluentui/tree/@fluentui/react-migration-v0-v9_v9.6.32)